requirements.txt
dj-rest-auth==6.0.0
Django==4.2.7
django-allauth==64.0.0
django-comment-migrate==0.1.7
django-cors-headers==4.3.0
django-filter==23.3
django-huey==1.2.0
django-ranged-response==0.2.0
django-redis==5.4.0
django-rest-swagger==2.2.0
django-restql==0.15.3
django-simple-captcha==0.5.20
django-timezone-field==6.0.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
einops
huey==2.5.1
imageio
loguru
matplotlib
numpy<=1.26.2
opencv-python
pillow
PyJWT
pynvml
pytorch-msssim
psutil
PyYAML
scikit-image
scipy
seaborn
SQLAlchemy
torch==1.11.0
torchvision==0.12.0
basicsr>=1.4.2
drf_yasg
tqdm
waitress